+++ .env
@@ -1,3 +1,3 @@
-VITE_APP_BASE_URL=https://localhost:7124/api
+VITE_APP_BASE_URL=https://api.drdentist.me/api
 
 VITE_APP_RECAPTCHA_SITE_KEY=6LdFTgMqAAAAAPMELcBY4rySQ1UpT5ZcXvVcj-_J
